Junk-Codes/Python/Voice-Recoder/pyproject.toml
[project]
name = "voice-recoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scipy-io-wavfile",
    "sounddevice",
]
